Bugfix: the behavior of the DISTINCT keyword operates differently in postgresql-4.x
It is entirely possible that the behavior I relied on previously was a bug that has been resolved.
Either way, I am now following what I read here:
- http://www.postgresql.org/docs/9.4/static/sql-select.html#SQL-DISTINCT

This changes the behavior of the distinct variable.
Instead of accepting a string, accept an array with the keys: 'field' and 'alias'.

This is a bugfix for: 544b7ca Feature: support DISTINCT(field.name) in addition to global DISTINCT