Update: redesign unrestricted.php to perform better, provide more specific error codes, and handle if_none_match (and similar) http headers
SSL mode is disabled per default for public_user, this immediately reduces the connection to the database by about 20ms.
- This further decreases performance for larger files based on number of packets.
  - For an ~1.3mB file, it previously took ~350ms and it now takes ~220ms.

Provide "406 Not Acceptable" error code to handle cases where invalid match data is provided (such as an etag that is invalid).

The "If None Matched" and similar functionality was never really tested until now.
- Get the code working.
- Testing it via browsers like firefox requires accessing about:config:
  - Set: browser.cache.memory.enable = TRUE
  - Set: browser.cache.check_doc_frequency = 1
  - Note: I have trouble getting browser.cache.disk.enable to work, so I turned mine off and only used memory.

For chrome to work, I more than likely need to make sure "Cache-Control:" is configured such that:
- 'no-cache' is not set.
- 'must-revalidate' is set.
- Consider supplying a 'max-age' directive, set it to something like 32 seconds.