Update: redesign killfile by hiding the nodes in a separate table (loses undelete)
This improves the performance by moving the killfile to a separate table.
This prevents having to merge or join the killfile table on every single select.

The undelete support is generally removed (but can be done via manual SQL queries).

This requires removing all of the killfile fields from views.

This is done in a compatible way such that the old database structure is preserved.