Update: cycle2 works properly when there is only 1 transition
If cycle2 is set to use more than one transition by placing a transition on a per-image basis, then "white box" issue happens.
This is a result of cycle2's design such that different transitions alter the same image at the same time causing undefined behavior.

After switching to a single transition, the cycle2 settings needed to be altered to prevent bugs created by some of the settings.
- This mostly happened with the slide-navigation.