Update: import theme redesign branch changes for mcneese_www
The content will only appear if and only if the appropriate sub-theme is defined.
Setting the sub-theme requires explicit access so it should be safe to merge this into the mainline.

Doing this helps prevent separation between branches over time given how the redesign has been sitting untouched for quite a long time.