Bugfix: use htmlspecialchars() in addition to strip_tags()
The previous commit "Bugfix: replace check_plain() with strip_tags() for appropriate rss feed tags" is not complete.
The check_plain() function is for html markup and could produce non-xml compliant sanitization.
- Removing entirely is not quite the solution, instead it should have been replaced with the appropriate htmlspecialchars() or htmlentity() function.
  - I am uncertain which is more appropriate: htmlspecialchars() or htmlentity().

Try to avoid using CDATA as much as possible. There also does not seem to be a sanitization function to protect against CDATA escapes.
- I will likely look into removing CDATA usage altogether to prevent situations where ]]> is used.
- It may be possible that ]]> cannot happen because the > should be escaped to &gt;.