Update: change behavior of news and lagniappe rss feeds
The news and lagniapped already have overlap and there needs to be a generic 'frontpage' feed.
Using 'news' instead of 'frontpage' sounds better, so change the behavior of the 'news' feed to contain all front page content.
The 'lagniappe' feed must then take both the highlight and the event content in the same way the 'news' feed used to.