Bugfix: fix misuse of CSS, provide images in theme, and remove wwwdev2 references as necessary
The sub-theme should always use .node-theme-page-731  and not affect the main mcneese theme.
The images added by a css should be provided within the theme itself and not outside the theme (no guarantee the files exist outside).
wwwdev2 was created as a copy of wwwdev, if the custom background colors needed to be changed, then simply remove the wwwdev2 specific stuff instead of overriding the override!

Note: there are uses of !important that very likely should be removed.
- The proper use of the .node-theme-page-731 should cause an override anyway, but if that is not enough, then using by id (aka: #example) will have override ability as well.