Bugfix: strip out duplicate headings in rss feeds when <h1> is present
This can easily be observed with HR Job Opportunities.

The <h2> tag is added for accessibiltiy reasons, but the hr jobs supply a custom h1 tag and the order gets changed.
This results in two page titles, the default hidden <h2> tag and the additional <h1> tag.

When the <h1> tag is present, remove the additional/redudant hidden <h2> tag so that it does not get displayed in an RSS feed.