Bugfix: check file size before calling getimagesize() and hide printed message on execute of getimagesize()
getimagesize() likes to produce error messages instead of simply returning something (such as NULL or FALSE) on failure.
The code already checks for this, so if NULL or FALSE is returned, it knows that the file does not exist and does not attempt to process the image results.

Therefore, there is no problem here, so do not report a problem!
The easiest way to prevent this is to wrap the function in an @..

There also seems to be a suggestion on php.org's documentation for "Read Error!" bugs by first checking with filesize() before calling getimagesize().
I do not know if filesize() needs to also have @.
To play it safe, I am not adding @ to filesize() at this time so that an errors can be exposed for investigation.

see: http://php.net/manual/en/function.getimagesize.php
see: http://php.net/manual/en/function.filesize.php