Bugfix: multiple template inputs with the same only got processed once, leaving all other uses to be undefined
Only one usage of a template input with a given name would be processed.
This is because of the logic in the code caused multiple ti's to overwrite existing ti's.
Only the last ti would be processed and expanded.

The fix is to store the list of ti's by names as an array of objects instead of a single object per name.