Security: revert drupals 'ajax security fix', it breaks functionality for a client-side securty bug
Client Side security is not relevant to the site as all ajax data should be treated and handled as untrusted already.
This behavior is bad practice and fixes a single problem that is client side that can still be exploited via any method possible (because client side is client side!)

This so called 'fix' breaks behavior, does the opposite of what the code implies (in regards to the clean urls change), and exposes unecessary information to the client side that does nothing to improve security.
I consider their security fix a security issue.