Bugfix: use watchdog() to handle errors instead of a custom function
This is the proper drupal way and if the user wanted to display the error message, there is always a drupal_set_message().
This does not add drupal_set_message() as there is no reason to confuse/scare the user about this bug.
Let the logs take care of the error messages.

ticket: www-1266