Bugfix: image width and height should not be 0, and make sure extension is preserved
The getimagesize() function returned a valid data object, but failed to properly detect the width and height, setting both to 0 for a specific .ico file.
Take advantage of imagemagicks tools to determine the size without losing existing data.

Make extra effort to preserve the extension and mime type.

ticket: www-1266