Bugfix: rewrite imagemagicks tempname handling so that extensions are added
The function drupal_tempnam() auto-created the file, so if an extension is going to be added, then delete that tempfile and add the extension.

By adding the extension, badly designed file formats that lack headers, such as ICO, will be detectable by imagemagick via the filename extension.

Both the _imagemagick_convert() and image_imagemagick_get_info() functions need to have this change applied.

ticket: www-1266