Security: restrict file style generation to specific urls for dbr and dbu files
Previously image styles would be generated for every requested url under the appropriate path.
This can easily be abused.
As the image style urls are generated, there is no reason to support different filenames.

Provide 404 for all files that are not the correct name.
Provide 404 for all url paths that have unnecessary/extra arguments.