Update: increase default cache lifetime for ntas_widget and decrease socket open time
Waiting 8 seconds is too long for a page, set it to 1.
This was observed when the network off campus was not working, took 8 seconds every time the police page was loaded.

This needs to be fixed by adding a cache counter that counts the number of failed attempts within, say 1 minute and then stops trying to load until some time later.