Regression: $menu_total is modified but not used
As mentioned in the bugfix, menu total is not used and I need to keep an eye out for regressions.
The variable was not used this entire time so do not use it.
Comments added to make note of this situation.

commit: 77dd48f4b98f6de466f9798aaf9c61c1f00a8f74