Update: attempt to re-connect persistent connections if the connection is lost
The following errors were observed following a database update (restarting of the database server).
  PHP Notice:  pg_query(): Cannot set connection to blocking mode in unrestricted.php on line 507
  PHP Notice:  Unknown: Cannot set connection to blocking mode in Unknown on line 0

I managed to see this in the browser, which returned a white page displaying the 503 error code.
The re-occurance of the error was somewhat erratic but the error eventually went away on its own.

In an attempt to prevent this behavior, check the persistent connection status before executing the query.
If the connection is consider bad then attempt to reconnect, forcing a new persistent connection.
This action is performed only one time.

I cannot re-produce the problem, even when restarting the database or restarting the webserver.

This configuration option is worth investigating:
- pgsql.auto_reset_persistent = Off

I just do not trust it due to very poor documentation on the PHP website.
- It sounds straight forward, but there are no described details as to how it works.