Update: allow for manually specifying watchdog log limit
Change the select list to a textfield so that the log entry can be manually specified.
This grants the administrator full control over their log size limits.

For security, the cron job that maintains this limit will force the specified value to be an integer.
If the value is not numeric, then a small default is used.