Bugfix: rss feed cache flooding database logs with errors and not using cache
The database log keeps showing messages such as:
  ERROR:  duplicate key value violates unique constraint "cache_page_pkey"
  DETAIL:  Key (cid)=(feed:rss/feed/group/717/4e6ff97b2be1e756537ccf5c04d952b5e438217d64ea86ff72d54e44853dd357) already exists.
  STATEMENT:  INSERT INTO cache_page (cid, serialized, created, expire, data) VALUES ('feed:rss/feed/group/717/4e6ff97b2be1e756537ccf5c04d952b5e438217d64ea86ff72d54e44853dd357', '1', '1464095719', '1464096920', '\x4e3b')

This is caused by drupal not doing tests to see if a a row exists perform inserting to the cache table.
To solve this, a cache_get is performed using the checksum and if the checksum already exists, do not save.

This revealed that the caching was also including the timestamp as a result of the generated page.
- This means that every single request has a different checksum due to the timestamp changing, unless requests happen within the same second (resulting in above errors).
- Use the database output instead of the generated ical page to generate the checksum so that timestamps are not used.