Bugfix: onConflictUpdate for caching is more complex than originally expected
When telling postgresql 9.5+ to perform an "on conflict update" action (aka: an upsert) they update syntax is far more complex than I thought.
The previous testing did not show any immediate errors and I did not notice that it was not working properly.

This rewrites the related code to handle the additional arguments for upsert actions as defined by the upstream documentation.
This is not complete support for all of the options.
In particular, I tried to implement the WHERE support but the amount of changes turned out to be more work than is worth.

The primary purpose of this is to solve a serious problem with drupals cache_set() function.
The functionality provided is only sufficient enough to perform the cache_set() operations.

see: https://www.postgresql.org/docs/current/static/sql-insert.html#SQL-ON-CONFLICT
see: https://www.postgresql.org/about/news/1636/
see: https://www.depesz.com/2015/05/10/waiting-for-9-5-add-support-for-insert-on-conflict-do-nothingupdate/
see: http://michael.otacoo.com/postgresql-2/postgres-9-5-feature-highlight-upsert/
see: http://stackoverflow.com/questions/17267417/how-to-upsert-merge-insert-on-duplicate-update-in-postgresql
see: https://www.compose.io/articles/postgresqls-future-is-looking-upsert/
see: http://pgeoghegan.blogspot.com/2015/10/avoid-naming-constraint-directly-when.html