Bugfix: run the cache set operation within a transaction
The cache operations need to be locked, so lock the cache table from updating until the operations are complete.
This prevents concurrency issues that continue to happen.

This will have a performance penalty in terms of updating the cache.
This should have a stability improvement as well as an performance improvement in terms of error logs writing to the disk.