Bugfix: page break webform component causes accessibility issues and interferences with presentation
This was discovered while resolving ticket www-1443.

An H2 tag gets created by page break and there are no controls over it.
This changes the behavior to allow for changing the heading type (from h2 to h6, div, or span) or not display at all (new default).
This also adds the ability to specify custom class names on the page break title.

ticket: www-1443