Revert "Update: make RETURNING optional for on conflict do update (upsert) for performance reasons"
This reverts commit 99f09541fd12b577dc4b2192a40419ffdc472320.

Something weird happened and this commit disappeared from development branches.
I am rebuilding the repositories from master and I need to remove this commit in an attempt to fix this issue.
I will likely need to reapply it in some manner (at the least review it again to see if it is neccessary).