Bugfix: prevent Internal Server Errors by gracefully handling database errors in file_db
The file database module was throwing errors to the caller and drupal core wouldpresent an internal server error.
This can be avoided because if the file database is not accessed, only loaded files do not display.
The rest of the page (including any cached files) will still appear.

Instead of blocking the entire process, make sure that when an error happens report it and then continue execution.