Bugfix: Some conditionals are not working on the Conflic of Interest webform
This may cause problems with javascript and 'emails'.
That said, javascript should not have anything to do with e-mails and this code removed by this commit was there to ensure that javascript does nothing with e-mails.
Removing that code because it is somehow causing problems with conditionals should therefore be safe enough.

ticket: www-1504