Update: commit changes
Patrick has not commented on any of this.
The code needs to be sanitized and made compliant.
For now, just commit the changes so that it can be easier to come back and fix them.
Having received no information from Patrick, I have no idea on whether or not this should be committed.