Revert "Update: add <base> markup tag to hint at relative url absolute base url (to core theme)"
This reverts commit 0fd3095a0bd43104b852203bb4cb681d12ed9c0c.

For some unknown reason, this triggers the media modules image browser to load itself twice and then load the frontpage of www.
For some unknown reason, this does not happen when defined inside of the sub-theme.

Revertting this is a workaround as it does not solve the problem.
The problem is somewhere in media modules code because the <base> tag only informs about where relative paths originate.

ticket: www-1553