Bugfix: wsod caused by bad use of the same variable and array_walk
Use a foreach loop instead of an inline function passed to array_walk().
I had to guess as to which "$value" should be which "$value".
I did some testing and never noticed this function being called so I cannot guarantee my interpretation is correct.

I also renamed additional uses of $value so as to not be as confusing (and dangerous).
If $value was passed as an object, it would be a reference and this could result in the overwrite of the callers $value (unexpectedly).
Many drupal/PHP security issues have exploited this specific type of problem.

ticket: www-1559