Regression: search block form not accepting url argument
However drupal handles the url argument, it is failing to pass the search url argument to the form.
Apparently, the process requires the #name attribute to be explicity set to search_box_form.

This ends up not being accessible because the "for" attribute on the label tag should also point to the "name" attribute.
Hopefully the fact that the id attribute is sufficient to pass accessibility.

I've looked at the search_block_form drupal code and found no obvious way to properly solve this.
For now, the only solution is to revert/remove this regression and maintain the possible accessibility violation.

ticket: www-1611