Revert "Bugfix: prevent error messages by not processing eid conditionals in js (which they dont apply to)"
This reverts commit c5840e1a5473b27f191a007fa4d32eb3523e2353.

This seems to be a fix for older version of webform.
The current version must have changed its structure, breaking this fix.

Either the fix no longer needs to be applied or it needs to be re-created in accordance to the new code structure.
First thing to do is simply to revert this change and see what happens.

ticket: www-1683